# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Vdoc API
  version: 0.1.0
  description: Machine-readable contract for the Vdoc v0.1 HTTP API. Projects own typed Documents, including OpenAPI and Markdown. REST handlers return HTTP 200 with the semantic result in the Vdoc envelope body.
servers:
  - url: /
security: []
...
paths:
...
  /api/v1/private/projects/{project_id}/members:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
    get:
      tags: [Projects]
      summary: List project members.
      operationId: listProjectMembers
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocListEnvelopeResponse'}
    post:
      tags: [Projects]
      summary: Add project member.
      operationId: addProjectMember
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
  /api/v1/private/projects/{project_id}/member-candidates:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
    get:
      tags: [Projects]
      summary: List active users eligible to join the project.
      description: Project Admin or SuperAdmin only. Existing active project members and disabled users are excluded.
      operationId: listProjectMemberCandidates
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocListEnvelopeResponse'}
  /api/v1/private/projects/{project_id}/members/{user_id}:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/UserID'}
    delete:
      tags: [Projects]
      summary: Remove project member.
      operationId: removeProjectMember
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
  /api/v1/private/projects/{project_id}/members/{user_id}/role:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/UserID'}
    patch:
      tags: [Projects]
      summary: Change project member role.
      operationId: patchProjectMemberRole
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
...
  /api/v1/private/projects/{project_id}/documents/{document_id}/drafts:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/DocumentID'}
    get:
      tags: [Drafts]
      summary: List document drafts.
      operationId: listDocumentDrafts
      security: [{JWTAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/BranchIDQuery'}
      responses:
        '200': {$ref: '#/components/responses/VdocListEnvelopeResponse'}
    post:
      tags: [Drafts]
      summary: Create document draft.
      operationId: createDocumentDraft
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
  /api/v1/private/projects/{project_id}/documents/{document_id}/drafts/{draft_id}:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/DocumentID'}
      - {$ref: '#/components/parameters/DraftID'}
    get:
      tags: [Drafts]
      summary: Get document draft.
      operationId: getDocumentDraft
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
    patch:
      tags: [Drafts]
      summary: Update document draft.
      operationId: updateDocumentDraft
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
...
  /api/v1/private/projects/{project_id}/documents/{document_id}/drafts/{draft_id}/submit:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/DocumentID'}
      - {$ref: '#/components/parameters/DraftID'}
    post:
      tags: [Drafts]
      summary: Submit document draft.
      operationId: submitDocumentDraft
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
...
  /api/v1/private/projects/{project_id}/documents/{document_id}/versions:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/DocumentID'}
    get:
      tags: [Versions]
      summary: List published document versions.
      operationId: listDocumentVersions
      security: [{JWTAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/BranchIDQuery'}
      responses:
        '200': {$ref: '#/components/responses/VdocListEnvelopeResponse'}
...
  /api/v1/private/mcp-tokens:
    get:
      tags: [MCP Tokens]
      summary: List MCP tokens.
      operationId: listMCPTokens
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocListEnvelopeResponse'}
    post:
      tags: [MCP Tokens]
      summary: Create MCP token.
      description: Returns a copyable token on create. The owner can reveal an active token again with the detail endpoint; list, revoked, and expired responses are redacted.
      operationId: createMCPToken
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
  /api/v1/private/mcp-tokens/{token_id}:
    parameters:
      - {$ref: '#/components/parameters/TokenID'}
    get:
      tags: [MCP Tokens]
      summary: Get MCP token.
      operationId: getMCPToken
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
  /api/v1/private/mcp-tokens/{token_id}/revoke:
    parameters:
      - {$ref: '#/components/parameters/TokenID'}
    post:
      tags: [MCP Tokens]
      summary: Revoke MCP token.
      operationId: revokeMCPToken
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
components:
  securitySchemes:
    JWTAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Raw JWT string issued by login or register.
    MCPTokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Raw MCP token string issued by the private MCP token APIs.
    DocumentShareAuth:
      type: apiKey
      in: header
      name: Authorization
      description: VdocShare scheme carrying the fragment-derived document-share capability. Protected links also require X-Vdoc-Share-Unlock after password exchange.
  parameters:
    BranchIDQuery:
      name: branch_id
      in: query
      required: false
      schema: {type: string, format: uuid}
      description: Filter drafts or versions to one document branch.
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: {type: string}
      description: Project ID.
    DocumentID:
      name: document_id
      in: path
      required: true
      schema: {type: string}
      description: Document ID.
...
    DraftID:
      name: draft_id
      in: path
      required: true
      schema: {type: string}
      description: Draft ID.
...
    TokenID:
      name: token_id
      in: path
      required: true
      schema: {type: string}
      description: MCP token ID.
    UserID:
      name: user_id
      in: path
      required: true
      schema: {type: string}
      description: User ID.
...
  responses:
    VdocEnvelopeResponse:
      description: HTTP 200 response with Vdoc envelope.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/VdocEnvelope'}
    VdocListEnvelopeResponse:
      description: HTTP 200 response with Vdoc envelope and total count.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/VdocListEnvelope'}
...
  schemas:
    VdocEnvelope:
      type: object
      properties:
        code: {type: integer}
        status: {type: string}
        description: {type: string}
        message: {type: string}
        trace_id: {type: string}
        timestamp: {type: integer}
        detail: {}
    VdocListEnvelope:
      allOf:
        - {$ref: '#/components/schemas/VdocEnvelope'}
        - type: object
          properties:
            total: {type: integer}
...